feat(swift): offer Google and Apple on sign-up, and a way back to guest (#2737)
Two gaps in the auth flow, both about paths that existed but could not be
reached.

Google and Sign in with Apple were only rendered on the sign-in screen.
The providers already work for new users — Better Auth creates the account
on the first OAuth callback — so tapping "Continue with Google" as a
brand-new user was always a sign-up. Only the discoverability was missing,
and the practical effect was that a new user's one visible path was email
plus password: the slowest option and the one most likely to be abandoned.

The block is extracted to `AuthProviderButtons` rather than copied, so the
two screens cannot drift. On sign-up it sits above the email form, since a
one-tap option is worth seeing before you start typing a password. The
caller owns `isLoading` and `error` because a provider sign-in should
disable the whole form, not just that section. A cancelled Apple sheet no
longer surfaces an error — backing out was a choice, not a failure.

Separately, opening either sign-in or sign-up was a one-way door: neither
had a route back to the welcome screen, which is the only entrance to
guest mode. Both now have a Back button.

Verified on the iPhone 17 Pro simulator (iOS 26.4): sign-up shows "Or
continue with" plus both providers, sign-in keeps them after the refactor,
and Back returns to the welcome screen on both.
